Don't use MTP* in call and invoice media data.

+++ Telegram/SourceFiles/history/history_item_reply_markup.h
@@ -47,7 +47,7 @@ struct HistoryMessageMarkupButton {
 		const QString &text,
 		const QByteArray &data = QByteArray(),
 		const QString &forwardText = QString(),
-		int32 buttonId = 0);
+		int64 buttonId = 0);
 
 	static HistoryMessageMarkupButton *Get(
 		not_null<Data::Session*> owner,
@@ -58,7 +58,7 @@ struct HistoryMessageMarkupButton {
 	Type type;
 	QString text, forwardText;
 	QByteArray data;
-	int32 buttonId = 0;
+	int64 buttonId = 0;
 	mutable mtpRequestId requestId = 0;
 
 };
